Move NonEmptySetWrapper to util namespace, plus address review comments
