+++ model/shared/src/main/scala/explore/model/util/NonEmptySetWrapper.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2021 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.model
+package explore.model.util
 
 import cats.data.NonEmptySet
 import cats.syntax.all._
@@ -13,38 +13,53 @@ final class NonEmptySetWrapper[IdSet, Id](self: IdSet, iso: Iso[IdSet, NonEmptyS
   private val selfSet = iso.get(self)
 
   // expose the NonEmptySet methods
+  @inline
   def ++(other: IdSet): IdSet = iso.reverseGet(selfSet ++ iso.get(other))
 
+  @inline
   def --(other: IdSet): SortedSet[Id] = selfSet -- iso.get(other)
 
+  @inline
   def add(id: Id): IdSet = iso.reverseGet(selfSet.add(id))
 
+  @inline
   def filter(f: Id => Boolean): SortedSet[Id] = selfSet.filter(f)
 
+  @inline
   def contains(id: Id): Boolean = selfSet.contains(id)
 
+  @inline
   def exists(f: Id => Boolean): Boolean = selfSet.exists(f)
 
+  @inline
   def forall(f: Id => Boolean): Boolean = selfSet.forall(f)
 
+  @inline
   def intersect(other: IdSet): SortedSet[Id] =
     selfSet.intersect(iso.get(other))
 
+  @inline
   def size: Long = selfSet.size
 
+  @inline
   def toList: List[Id] = selfSet.toList
 
   // some helper methods
+  @inline
   def remove(other: IdSet): Option[IdSet] = fromSet(this -- other)
 
+  @inline
   def removeUnsafe(other: IdSet): IdSet =
     iso.reverseGet(NonEmptySet.fromSetUnsafe(selfSet -- iso.get(other)))
 
+  @inline
   def removeOne(id: Id): Option[IdSet] = fromSet(selfSet - id)
 
+  @inline
   def filterOpt(f: Id => Boolean): Option[IdSet] = fromSet(selfSet.filter(f))
 
-  def hasIntersect(other: IdSet): Boolean = intersect(other).nonEmpty
+  @inline
+  def intersects(other: IdSet): Boolean = intersect(other).nonEmpty
 
   private def fromSet(set: SortedSet[Id]): Option[IdSet] =
     NonEmptySet.fromSet(set).map(iso.reverseGet)
